Extract MISP and osquery evidence helpers
