+++ control-plane/tests/test_phase28_external_evidence_boundary_refactor.py
@@ -13,6 +13,7 @@ if str(CONTROL_PLANE_ROOT) not in sys.path:
 
 
 from aegisops_control_plane.config import RuntimeConfig
+from aegisops_control_plane import external_evidence_boundary
 from aegisops_control_plane.service import AegisOpsControlPlaneService
 from postgres_test_support import make_store
 
@@ -257,6 +258,23 @@ class Phase28ExternalEvidenceBoundaryRefactorTests(unittest.TestCase):
                 self.assertFalse(
                     hasattr(AegisOpsControlPlaneService, helper_name),
                     f"{helper_name} should be owned by the external evidence boundary",
+                )
+
+    def test_misp_and_osquery_helpers_are_extracted_from_boundary_module(self) -> None:
+        extracted_helper_names = (
+            "_normalize_misp_indicator_type",
+            "_normalize_misp_indicator_value",
+            "_mapping_matches_misp_indicator",
+            "_container_explicitly_cites_misp_indicator",
+            "_require_explicit_misp_anchor_binding",
+            "_require_case_host_identifier",
+        )
+
+        for helper_name in extracted_helper_names:
+            with self.subTest(helper_name=helper_name):
+                self.assertFalse(
+                    hasattr(external_evidence_boundary, helper_name),
+                    f"{helper_name} should be owned by focused external evidence modules",
                 )
